Downgrade eslint-plugin-react-hooks
The latest release of eslint-plugin-react-hooks incorrectly flags uses of hooks as occurring in a loop if a component function body contains a loop.

Downgrading to version 5.0.0 is a workaround until the React team publishes a release with the fix in https://github.com/facebook/react/pull/28714 (merged but not yet released).

See https://github.com/facebook/react/issues/31687